backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
numpy==2.2.5

scikit-learn>=1.4.2

# TensorFlow 2.15 has no wheel for newer Python versions; use 2.20 which pip shows available
# (if you require 2.15, use Python 3.11/3.10). See notes below.
tensorflow==2.20.0
xgboost==2.0.3
requests==2.31.0
pymongo==4.6.0
